+++ src/features/posts/constants/post-form-field-name.ts
@@ -1,0 +1,7 @@
+export const POST_FORM_FIELD_NAME = {
+  title: 'title',
+  content: 'content',
+  isPhotoVisibleToDesigner: 'isPhotoVisibleToDesigner',
+  imageFiles: 'imageFiles',
+  imageUrls: 'imageUrls',
+} as const;
